Rename RecoilValue core files (#322)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/322

Rename the core files for RecoilValues for clarity.
* `Recoil_RecoilValue.js` - contains the actual RecoilValue and related classes
* `Recoil_RecoilValueInterface.js` - contains the the interface for using RecoilValues

Reviewed By: aaronabramov

Differential Revision: D22029908

fbshipit-source-id: dc66870763b88096a49535064cd1519cf7c473f5
