+++ src/recoil_values/__tests__/Recoil_atomFamily-test.js
@@ -17,7 +17,7 @@ const {act} = require('ReactTestUtils');
 const {
   getRecoilValueAsLoadable,
   setRecoilValue,
-} = require('../../core/Recoil_RecoilValue');
+} = require('../../core/Recoil_RecoilValueInterface');
 const {
   useRecoilState,
   useRecoilValue,
