+++ src/recoil_values/__tests__/Recoil_constSelector-test.js
@@ -10,7 +10,9 @@
  */
 'use strict';
 
-const {getRecoilValueAsLoadable} = require('../../core/Recoil_RecoilValue');
+const {
+  getRecoilValueAsLoadable,
+} = require('../../core/Recoil_RecoilValueInterface');
 const {makeStore} = require('../../testing/Recoil_TestingUtils');
 const constSelector = require('../Recoil_constSelector');
