+++ src/recoil_values/__tests__/Recoil_selectorFamily-test.js
@@ -24,7 +24,7 @@ const selectorFamily = require('../Recoil_selectorFamily');
 const {
   getRecoilValueAsLoadable,
   setRecoilValue,
-} = require('../../core/Recoil_RecoilValue');
+} = require('../../core/Recoil_RecoilValueInterface');
 const {makeStore} = require('../../testing/Recoil_TestingUtils');
 
 let store;
